+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added editionID validation to editions API
 - Updated README with steps on retrieving local-compose.yaml file and credentials
 - Added more logging to proxy API
+- Refactor OCLC Catalog Manager
 ## Fixed
 - Resolved the format of fulfill endpoints in UofM manifests
 - Added additional logging to the editions endpoint to debug